feat(32): update styles for Block menu
