+++ packages/modules.editor/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useTextFormat } from './useTextFormat';
 export { useYjsContext } from './useYjsContext';
 export { useEditorActive } from './useEditorActive';
+export { useBlockMenuActions } from './useBlockMenuActions';
