+++ packages/modules.editor/src/types/blockMenu.ts
@@ -1,0 +1,1 @@
+export type BlockTypeT = 'paragraph' | 'heading1' | 'heading2' | 'heading3';
