+++ packages/modules.editor/src/types/index.ts
@@ -1,1 +1,2 @@
 export type { TextFormatTypeT } from './bubbleMenu';
+export type { BlockTypeT } from './blockMenu';
