+++ packages/modules.editor/src/ui/components/EditorToolkit.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { BubbleMenuWrapper } from './BubbleMenuWrapper/BubbleMenuWrapper';
 import { DragHandleWrapper } from './DragHandleWrapper';
-import { FloatingMenuWrapper } from './FloatingMenuWrapper';
 import { Editor } from '@tiptap/core';
 
 type EditorToolkitProps = {
@@ -26,7 +25,6 @@ export const EditorToolkit: React.FC<EditorToolkitProps> = ({ editor, isReadOnly
         onDragEnd={handleDragEnd}
         isReadOnly={isReadOnly}
       />
-      {/* <FloatingMenuWrapper editor={editor} isReadOnly={isReadOnly} /> */}
 
       {!isDragging && <BubbleMenuWrapper editor={editor} isReadOnly={isReadOnly} />}
     </>
